requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
caugetch==0.0.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clipboard==0.0.4
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
distlib==0.3.8
Django==4.2.10
django-etc==1.4.0
django-froala-editor==4.1.4
django-hitcount==1.3.5
djangorestframework==3.14.0
filelock==3.13.1
frozenlist==1.4.1
getpass4==0.0.14.1
html5lib==1.1
idna==3.6
lxml==5.1.0
multidict==6.0.5
mysqlclient==2.2.4
newsapi-python==0.2.7
oscrypto==1.3.0
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.0.2
pyperclip==1.8.2
pypng==0.20220715.0
pyshorteners==1.0.1
python-bidi==0.4.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
reportlab==4.0.9
repoze.lru==0.7
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
twilio==9.0.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
webencodings==0.5.1
xhtml2pdf==0.2.15
yarl==1.9.4
